# src/main/java/com/backend/ticketing/dto/VehicleDTO.java
package com.backend.ticketing.dto;

import com.backend.ticketing.enums.VehicleType;

public class VehicleDTO {

    private String registrationNumber;
    private VehicleType type;
    private int capacity;

    public String getRegistrationNumber() {
        return registrationNumber;
    }

    public void setRegistrationNumber(String registrationNumber) {
        this.registrationNumber = registrationNumber;
    }

    public VehicleType getType() {
        return type;
    }

    public void setType(VehicleType type) {
        this.type = type;
    }

    public int getCapacity() {
        return capacity;
    }

    public void setCapacity(int capacity) {
        this.capacity = capacity;
    }
}