# src/main/resources/META-INF/openapi.yaml
openapi: 3.0.1
info:
  title: Unicard Ticketing API
  description: REST API for managing public transport ticketing
  version: "1.0.0"
paths:
  /vehicles:
    post:
      summary: Register a new vehicle
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Vehicle'
      responses:
        "201":
          description: Vehicle created
    get:
      summary: List all registered vehicles
      responses:
        "200":
          description: List of vehicles
...
  /tickets/vehicle/{reg}:
    get:
      summary: List tickets for a given vehicle
      parameters:
        - name: reg
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of tickets
...
components:
  schemas:
    Vehicle:
      type: object
      properties:
        registrationNumber:
          type: string
        type:
          type: string
        passengerCapacity:
          type: integer
...